WIP html pass values to backend
a text box will take user input and the user input will be compare with a dictionary that contains all 172 ip to verify if the user input is a real ip. If the user input is in real ip list, the script will trigger a function/script to purge the content on that ip address.